WIP: Moves field mapping code to its own files; finishes validation service enum
