+++ pkg/v1/tkg/web/src/app/views/landing/wizard/shared/constants/validation.constants.ts
@@ -104,9 +104,9 @@ export enum SimpleValidator {
     IS_VALID_CLUSTER_NAME,
     IS_VALID_FQDN,
     IS_VALID_FQDN_OR_IP,
+    IS_VALID_FQDN_OR_IP_HTTPS,
     IS_VALID_FQDN_OR_IP_LIST,
     IS_VALID_FQDN_OR_IPV6,
-    IS_VALID_FQDN_OR_IP_HTTPS,
     IS_VALID_FQDN_OR_IPV6_HTTPS,
     IS_VALID_IP,
     IS_VALID_IP_LIST,
@@ -115,7 +115,6 @@ export enum SimpleValidator {
     IS_VALID_LABEL_OR_ANNOTATION,
     IS_VALID_PORT,
     IS_VALID_RESOURCE_GROUP_NAME,
-    IS_VALID_SUBNET,
     NO_WHITE_SPACE,
     NO_TRAILING_SLASH
 }
